refactor(useNotifications): simplify syntax
